LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/pixel-manipulation/2d.imageData.get.zero-expected.txt

2d.imageData.get.zero
getImageData() throws INDEX_SIZE_ERR if size is zero
Actual output:

PASS getImageData() throws INDEX_SIZE_ERR if size is zero
